fix: correction get business is column bank

See merge request 123milhas/finances/api!36

+++ app/Http/Requests/PutBankRequest.php
@@ -18,7 +18,7 @@ class PutBankRequest extends FormRequest
             'cnpj' => 'max:45',
             'cnab400' => 'boolean',
             'cnab240' => 'boolean',
-            'bank_code' => 'integer',
+            'bank_code' => 'numeric',
         ];
     }
 }
